requirements.txt
azure-common==1.1.28
azure-core==1.38.2
azure-identity==1.25.2
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.6.0
azure-mgmt-resource==25.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
idna==3.11
isodate==0.7.2
msal==1.35.0
msal-extensions==1.3.1
numpy==2.4.2
pandas==3.0.1
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
